# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "GuestIn": {
        "description": "Payload de creation/mise a jour d'une fiche client.\n\nTous les champs sont optionnels sauf nom/prenom : une reception peut\nouvrir une fiche minimale a l'arrivee et la completer plus tard, plutot\nque de bloquer le check-in en attendant une adresse complete.",
        "properties": {
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "birth_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Date"
          },
          "birth_place": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Place"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "company_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Id"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "credit_limit": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credit Limit"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "first_name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "First Name",
            "type": "string"
          },
          "gender": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gender"
          },
          "id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "UUID v7 genere par la tablette ; absent = genere par le serveur",
            "title": "Id"
          },
          "id_document_expiry": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id Document Expiry"
          },
          "id_document_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id Document Number"
          },
          "id_document_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IdDocumentType"
              },
              {
                "type": "null"
              }
            ]
          },
          "is_vip": {
            "default": false,
            "title": "Is Vip",
            "type": "boolean"
          },
          "last_name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Last Name",
            "type": "string"
          },
          "marketing_consent": {
            "default": false,
            "title": "Marketing Consent",
            "type": "boolean"
          },
          "nationality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nationality"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "phone_alt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone Alt"
          },
          "postal_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Postal Code"
          },
          "preferences": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferences"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "first_name",
          "last_name"
        ],
        "title": "GuestIn",
        "type": "object"
      },
      "GuestOut": {
        "properties": {
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "birth_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birth Date"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "company_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Id"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "credit_limit": {
            "title": "Credit Limit",
            "type": "integer"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "first_name": {
            "title": "First Name",
            "type": "string"
          },
          "full_name": {
            "title": "Full Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "id_document_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id Document Number"
          },
          "id_document_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IdDocumentType"
              },
              {
                "type": "null"
              }
            ]
          },
          "is_blacklisted": {
            "title": "Is Blacklisted",
            "type": "boolean"
          },
          "is_vip": {
            "title": "Is Vip",
            "type": "boolean"
          },
          "last_name": {
            "title": "Last Name",
            "type": "string"
          },
          "nationality": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nationality"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "id",
          "code",
          "full_name",
          "title",
          "first_name",
          "last_name",
          "birth_date",
          "nationality",
          "id_document_type",
          "id_document_number",
          "email",
          "phone",
          "address",
          "city",
          "country",
          "company_id",
          "is_vip",
          "is_blacklisted",
          "credit_limit",
          "notes"
        ],
        "title": "GuestOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IdDocumentType": {
        "enum": [
          "ID_CARD",
          "PASSPORT",
          "DRIVING_LICENSE",
          "RESIDENCE_PERMIT",
          "OTHER"
        ],
        "title": "IdDocumentType",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Atrium",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/guests": {
      "get": {
        "operationId": "list_guests_api_v1_guests_get",
        "parameters": [
          {
            "description": "Recherche nom, telephone ou email",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Recherche nom, telephone ou email",
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GuestOut"
                  },
                  "title": "Response List Guests Api V1 Guests Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Guests",
        "tags": [
          "clients"
        ]
      },
      "post": {
        "description": "Cree une fiche client, ou la met a jour si la tablette renvoie le meme id.\n\nUne tablette qui perd le reseau apres l'envoi ne sait pas si l'ecriture est\npassee et la renvoie : sans upsert, chaque coupure creerait un second\nclient. Une seule instruction `INSERT ... ON CONFLICT (id) DO UPDATE`,\nrestreinte a l'hotel de l'utilisateur : un id qui appartient a un autre\netablissement ne met rien a jour et repond 404. 201 a la creation, 200\nsur un renvoi. Le code client (CLI-...) n'est attribue qu'a la creation.",
        "operationId": "create_guest_api_v1_guests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuestIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Fiche deja connue (meme id) : mise a jour"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuestOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Guest",
        "tags": [
          "clients"
        ]
      }
    },
    "/api/v1/guests/{guest_id}": {
      "get": {
        "operationId": "get_guest_api_v1_guests__guest_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "guest_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Guest Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuestOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Guest",
        "tags": [
          "clients"
        ]
      },
      "patch": {
        "operationId": "update_guest_api_v1_guests__guest_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "guest_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Guest Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuestIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuestOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Guest",
        "tags": [
          "clients"
        ]
      }
    },
...
  }
}